Move app init to a single place
