--- anyway/__init__.py
@@ -1,1 +1,0 @@
-from .flask_app import app
