+++ anyway/app.py
@@ -1,0 +1,6 @@
+from flask_sqlalchemy import SQLAlchemy
+
+from anyway.utilities import init_flask
+
+app = init_flask()
+db = SQLAlchemy(app)
