+++ anyway/parsers/rsa.py
@@ -1,15 +1,13 @@
 # -*- coding: utf-8 -*-
 import json
+
+from anyway.app import db
 from ..constants import CONST
 from ..models import AccidentMarker
-from ..utilities import init_flask
 from .utils import batch_iterator
-from flask_sqlalchemy import SQLAlchemy
 from openpyxl import load_workbook
 from dateutil import parser
 
-app = init_flask()
-db = SQLAlchemy(app)
 
 def _iter_rows(filename):
     workbook = load_workbook(filename, read_only=True)
@@ -58,9 +56,6 @@ def _iter_rows(filename):
 
 
 def parse(filename):
-    app = init_flask()
-    db = SQLAlchemy(app)
-
     for batch in batch_iterator(_iter_rows(filename), batch_size=50):
         db.session.bulk_insert_mappings(AccidentMarker, batch)
         db.session.commit()
